+++ html/pfappserver/root/static.alt/src/globals/configuration/pfConfigurationBillingTiers.js
@@ -7,7 +7,7 @@ import {
   pfConfigurationListColumns,
   pfConfigurationListFields,
   pfConfigurationViewFields
-} from '@/globals/pfConfiguration'
+} from '@/globals/configuration/pfConfiguration'
 import {
   and,
   not,
