+++ html/pfappserver/root/static.alt/src/globals/configuration/pfConfigurationProfiling.js
@@ -4,7 +4,7 @@ import pfFormToggle from '@/components/pfFormToggle'
 import {
   pfConfigurationListColumns,
   pfConfigurationListFields
-} from '@/globals/pfConfiguration'
+} from '@/globals/configuration/pfConfiguration'
 import {
   and,
   isFQDN,
