+++ html/pfappserver/root/static.alt/src/globals/configuration/pfConfigurationRealms.js
@@ -6,7 +6,7 @@ import pfFormToggle from '@/components/pfFormToggle'
 import {
   pfConfigurationListColumns,
   pfConfigurationListFields
-} from '@/globals/pfConfiguration'
+} from '@/globals/configuration/pfConfiguration'
 import {
   and,
   not,
