+++ html/pfappserver/root/static.alt/src/views/Configuration/_components/ProfilingCombinationsList.vue
@@ -29,7 +29,7 @@ import pfEmptyTable from '@/components/pfEmptyTable'
 import {
   pfConfigurationProfilingCombinationsListColumns as columns,
   pfConfigurationProfilingCombinationsListFields as fields
-} from '@/globals/pfConfigurationProfiling'
+} from '@/globals/configuration/pfConfigurationProfiling'
 
 export default {
   name: 'ProfilingCombinationsList',
