+++ html/pfappserver/root/static.alt/src/views/Configuration/_components/ProfilingDhcpVendorsList.vue
@@ -29,7 +29,7 @@ import pfEmptyTable from '@/components/pfEmptyTable'
 import {
   pfConfigurationProfilingDhcpVendorsListColumns as columns,
   pfConfigurationProfilingDhcpVendorsListFields as fields
-} from '@/globals/pfConfigurationProfiling'
+} from '@/globals/configuration/pfConfigurationProfiling'
 
 export default {
   name: 'ProfilingDhcpVendorsList',
