+++ html/pfappserver/root/static.alt/src/views/Configuration/_components/ProfilingDhcpv6EnterprisesList.vue
@@ -29,7 +29,7 @@ import pfEmptyTable from '@/components/pfEmptyTable'
 import {
   pfConfigurationProfilingDhcpv6EnterprisesListColumns as columns,
   pfConfigurationProfilingDhcpv6EnterprisesListFields as fields
-} from '@/globals/pfConfigurationProfiling'
+} from '@/globals/configuration/pfConfigurationProfiling'
 
 export default {
   name: 'ProfilingDhcpv6EnterprisesList',
