+++ html/pfappserver/root/static.alt/src/views/Configuration/_components/ProfilingMacVendorsList.vue
@@ -29,7 +29,7 @@ import pfEmptyTable from '@/components/pfEmptyTable'
 import {
   pfConfigurationProfilingMacVendorsListColumns as columns,
   pfConfigurationProfilingMacVendorsListFields as fields
-} from '@/globals/pfConfigurationProfiling'
+} from '@/globals/configuration/pfConfigurationProfiling'
 
 export default {
   name: 'ProfilingMacVendorsList',
