+++ html/pfappserver/root/static.alt/src/views/Configuration/_components/ProfilingUserAgentsList.vue
@@ -29,7 +29,7 @@ import pfEmptyTable from '@/components/pfEmptyTable'
 import {
   pfConfigurationProfilingUserAgentsListColumns as columns,
   pfConfigurationProfilingUserAgentsListFields as fields
-} from '@/globals/pfConfigurationProfiling'
+} from '@/globals/configuration/pfConfigurationProfiling'
 
 export default {
   name: 'ProfilingUserAgentsList',
